Option to skip checking the encryption key exists
Should be a very slight performance improvement, since the check is made
when creating a new instance of the encryption helper. If you know for
sure either the check has already been made, or that the key exists you
can set the third (optional) parameter to false, and it'll skip the
check.

Defaults to the safer option of checking.